pyproject.toml
[project]
name = "terminus-cli"
version = "0.1.0"
description="Access terminals from your computer, in your browser."
readme = {file = "README.md", content-type = "text/markdown" }
requires-python = ">= 3.10"
dependencies = [
    "click>=8.0.0",
    "fastapi[standard]",
    "python-dotenv",
    "rich>=13.7.0",
    "httpx>=0.24.0",
    "websockets>=10.0",
    "python-multipart>=0.0.5",
    "tomli>=2.0.1; python_version < '3.11'"  # For Python versions that don't have tomllib
]
license = {file = "LICENSE"}
authors = [{name = "Alex TYRODE", email = "alex.tyrode@outlook.fr"}]
classifiers=[
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

[project.scripts]
terminus = "cli.app:cli"

[project.urls]
homepage = "https://terminus.run"
github = "https://github.com/atyrode/terminus"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.setuptools]
packages = {find = {where = ["src"]}}
include-package-data = true

# Linting configuration
[tool.pylint.format]
max-line-length = 119

[tool.pylint.'MESSAGES CONTROL']
disable = [
    "too-many-arguments",
    "too-many-positional-arguments",
    "too-many-locals",
    "too-few-public-methods",
    "broad-exception-caught",
    "duplicate-code",
]

[tool.pycln]
all = true

[tool.isort]
profile = "black"
line_length = 119
remove_redundant_aliases = true
float_to_top = true
order_by_type = true
filter_files = true

[tool.black]
line-length = 119

# Test configuration
[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests/"]
python_files = "test_*.py"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
python_classes = "Test*"
python_functions = "test_*"

[tool.hatch.build.targets.wheel]
packages = ["src/"]
